ingestion-agent/requirements.txt

langgraph
langchain_openai
azure-identity
selenium
markdownify
azure-storage-blob
geopy
